% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBreakdown.R
\name{summaryBreakdown}
\alias{summaryBreakdown}
\title{Step 3-1: Get prediction breakdown for a single row of data}
\usage{
summaryBreakdown(xgb.breakdown, DMatrix, data.matrix, idx,
  type = "binary", threshold = 1e-04, .show.result = TRUE)
}
\arguments{
\item{xgb.breakdown}{A xgb.breakdown object returnd by explainPredictions.}

\item{DMatrix}{The DMatrix in which the row to be predicted is stored}

\item{data.matrix}{The matrix of data from which the DMatrix was built}

\item{idx}{The row number of the data to be explained}

\item{type}{The objective function of the model - either "binary" (for binary:logistic) or "regression" (for reg:linear)}

\item{threshold}{Default = 0.0001. The waterfall chart will group all variables with absolute impact less than the threshold into a variable called 'Other'}
}
\value{
a list
}
\description{
This function return a list of the feature impact breakdown for a single data row.
}
