% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallBreakdown.R
\name{waterfallBreakdown}
\alias{waterfallBreakdown}
\title{Step 3-2: Waterfall chart for prediction breakdow of a single row of data}
\usage{
waterfallBreakdown(summary.breakdown, limits = c(NA, NA))
}
\arguments{
\item{summary.breakdown}{A summary.breakdown object returnd by summaryBreakdown().}

\item{limits}{The limits of the y axis - for binary this is on logit scale (e.g. c(-3,3) would give a scale approximately from 0.04 to 0.96)
Step 3-1: Get prediction breakdown for a single row of data

This function return a list of the feature impact breakdown for a single data row.}
}
\value{
a ggplots object from waterfalls
}
\description{
This function plots as waterfall chart of the feature impact breakdown for a single data row.
}
